moving and rename keySettings to settingSettings
