--- Settings/AppSettings/RemoveKeySettings.cs
@@ -1,6 +1,0 @@
-﻿using Spectre.Console.Cli;
-
-namespace config.Settings.AppSettings;
-internal class RemoveKeySettings : CommandSettings
-{
-}
