--- Settings/AppSettings/UpdateKeySettings.cs
@@ -1,7 +1,0 @@
-﻿using Spectre.Console.Cli;
-
-namespace config.Settings.AppSettings;
-internal class UpdateKeySettings : CommandSettings
-{
-    
-}
