+++ Settings/Settings/RemoveSettingSettings.cs
@@ -1,0 +1,6 @@
+﻿using Spectre.Console.Cli;
+
+namespace config.Settings.Settings;
+internal class RemoveSettingSettings : CommandSettings
+{
+}
