+++ Settings/Settings/UpdateSettingSettings.cs
@@ -1,0 +1,7 @@
+﻿using Spectre.Console.Cli;
+
+namespace config.Settings.Settings;
+internal class UpdateSettingSettings : CommandSettings
+{
+
+}
